requirements.txt
pylint>=2.3
psycopg2-binary>=2.8
pandas>=0.24
boto3>=1.9
python-dotenv>=0.10
botocore>=1.12

